Move PromiseQueue to use-create-missing-menu-items as it is only used there
